upgrade delay update
